+++ next-cloudinary/src/components/CldImage/CldImage.tsx
@@ -1,15 +1,19 @@
-'use client';
-
-import { useState } from 'react';
-import Image from 'next/image';
+import React, { useState } from 'react';
+import Image, { ImageProps } from 'next/image';
 import { getTransformations } from '@cloudinary-util/util';
 import { transformationPlugins } from '@cloudinary-util/url-loader';
+import type { ImageOptions } from '@cloudinary-util/url-loader';
 
 import { pollForProcessingImage } from '../../lib/cloudinary';
 
 import { cloudinaryLoader } from '../../loaders/cloudinary-loader';
 
-const CldImage = props => {
+export type CldImageProps = Omit<ImageProps, 'src'> & ImageOptions & {
+  src: string;
+  preserveTransformations?: boolean;
+};
+
+const CldImage = (props: CldImageProps) => {
   const CLD_OPTIONS = [
     'deliveryType',
     'preserveTransformations'
@@ -26,34 +30,28 @@ const CldImage = props => {
 
   // Construct the base Image component props by filtering out Cloudinary-specific props
 
-  interface ImageProps {
-    alt: string;
-    blurDataURL?: string;
-    src: string;
-  }
-
-  const imageProps: ImageProps = {
+  const imageProps = {
     alt: props.alt,
     src: props.src
   };
 
-  Object.keys(props)
+  (Object.keys(props) as Array<keyof typeof props>)
     .filter(key => !CLD_OPTIONS.includes(key))
+    // @ts-expect-error
     .forEach(key => imageProps[key] = props[key]);
 
-  const defaultImgKey = Object.keys(imageProps).map(key => `${key}:${imageProps[key]}`).join(';');
+
+  const defaultImgKey = (Object.keys(imageProps) as Array<keyof typeof imageProps>).map(key => `${key}:${imageProps[key]}`).join(';');
   const [imgKey, setImgKey] = useState(defaultImgKey);
 
   // Construct Cloudinary-specific props by looking for values for any of the supported prop keys
 
-  interface CldOptions {
-    rawTransformations?: string[];
-  }
-
-  const cldOptions: CldOptions = {};
+  const cldOptions = {};
 
   CLD_OPTIONS.forEach(key => {
+    // @ts-expect-error
     if ( props[key] ) {
+      // @ts-expect-error
       cldOptions[key] = props[key] || undefined;
     }
   });
@@ -65,6 +63,7 @@ const CldImage = props => {
   if (props.preserveTransformations) {
     try {
       const transformations = getTransformations(props.src);
+      // @ts-expect-error
       cldOptions.rawTransformations = [...transformations.flat(), ...(props.rawTransformations || [])];
     } catch(e) {
       console.warn(`Failed to preserve transformations: ${(e as Error).message}`)
@@ -75,7 +74,11 @@ const CldImage = props => {
    * handleOnError
    */
 
-  async function handleOnError(options) {
+  interface HandleOnError {
+    target: any;
+  }
+
+  async function handleOnError(options: HandleOnError) {
     const result = await pollForProcessingImage({ src: options.target.src })
     if ( result ) {
       setImgKey(`${defaultImgKey};${Date.now()}`);