+++ next-cloudinary/src/components/CldUploadButton/CldUploadButton.tsx
@@ -1,4 +1,10 @@
-import CldUploadWidget from '../CldUploadWidget';
+import React from 'react';
+import CldUploadWidget, { CldUploadWidgetProps } from '../CldUploadWidget';
+
+export interface CldUploadButtonProps extends Omit<CldUploadWidgetProps, 'children'> {
+  children?: JSX.Element | string | Array<JSX.Element|string>;
+  onClick?: Function;
+}
 
 const CldUploadButton = ({
   children,
@@ -8,10 +14,9 @@ const CldUploadButton = ({
   signatureEndpoint,
   uploadPreset,
   ...props
-}) => {
+}: CldUploadButtonProps) => {
   return (
     <>
-      {/*  @ts-ignore */}
       <CldUploadWidget
         onUpload={onUpload}
         options={options}
@@ -19,7 +24,7 @@ const CldUploadButton = ({
         uploadPreset={uploadPreset}
       >
         {({ open }) => {
-          function handleOnClick(e) {
+          function handleOnClick(e: React.MouseEvent<HTMLButtonElement, MouseEvent>) {
             e.preventDefault();
 
             open();