+++ next-cloudinary/src/components/CldUploadButton/index.ts
@@ -1,0 +1,2 @@
+export { default } from './CldUploadButton';
+export type { CldUploadButtonProps } from './CldUploadButton';