+++ next-cloudinary/src/components/CldUploadWidget/index.ts
@@ -1,0 +1,2 @@
+export { default } from './CldUploadWidget';
+export type { CldUploadWidgetProps, CldUploadWidgetPropsOptions } from './CldUploadWidget';