+++ next-cloudinary/src/index.ts
@@ -1,6 +1,16 @@
 export { default as CldImage } from './components/CldImage';
+export type { CldImageProps } from './components/CldImage';
+
 export { default as CldOgImage } from './components/CldOgImage';
+export type { CldOgImageProps } from './components/CldOgImage';
+
 export { default as CldUploadButton } from './components/CldUploadButton';
+export type { CldUploadButtonProps } from './components/CldUploadButton';
+
 export { default as CldUploadWidget } from './components/CldUploadWidget';
+export type { CldUploadWidgetProps, CldUploadWidgetPropsOptions } from './components/CldUploadWidget';
+
 export { default as CldVideoPlayer } from './components/CldVideoPlayer';
+
 export { cloudinaryLoader } from './loaders/cloudinary-loader';
+export type { CloudinaryLoader, CloudinaryLoaderLoaderOptions, CloudinaryLoaderCldOptions } from './loaders/cloudinary-loader';