+++ next-cloudinary/tests/nextjs-app/app/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { CldImage } from '../../../';
+
+export default function Home() {
+  return <CldImage src="images/woman-headphones" width="100" height="100" alt="test" />
+}