+++ next-cloudinary/tests/nextjs-app/next.config.js
@@ -1,0 +1,11 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  env: {
+    NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME: process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME || 'test'
+  },
+  experimental: {
+    appDir: true,
+  },
+}
+
+module.exports = nextConfig