+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [2.3.2](https://github.com/colbyfayock/next-cloudinary/compare/v2.3.1...v2.3.2) (2023-02-20)
+
+
+### Bug Fixes
+
+* Fix Next 13 Use Client Directive ([#136](https://github.com/colbyfayock/next-cloudinary/issues/136)) ([9ee962d](https://github.com/colbyfayock/next-cloudinary/commit/9ee962d3db20a69482e2509123521f1422a9d318)), closes [#135](https://github.com/colbyfayock/next-cloudinary/issues/135)
+
+## [2.3.1](https://github.com/colbyfayock/next-cloudinary/compare/v2.3.0...v2.3.1) (2023-02-18)
+
+
+### Bug Fixes
+
+* Import CldImageProps from URL Loader ([#134](https://github.com/colbyfayock/next-cloudinary/issues/134)) ([e443b05](https://github.com/colbyfayock/next-cloudinary/commit/e443b0535f8daaba10fb5822622fa95930fb980d)), closes [#133](https://github.com/colbyfayock/next-cloudinary/issues/133)
+
+# [2.3.0](https://github.com/colbyfayock/next-cloudinary/compare/v2.2.0...v2.3.0) (2023-02-17)
+
+
+### Features
+
+* Migrate to tsup, Fix Types Exports ([#131](https://github.com/colbyfayock/next-cloudinary/issues/131)) ([c51d54f](https://github.com/colbyfayock/next-cloudinary/commit/c51d54fd499c8bfd62a00502d79be0ec458b7822)), closes [#130](https://github.com/colbyfayock/next-cloudinary/issues/130)
+
 # [2.2.0](https://github.com/colbyfayock/next-cloudinary/compare/v2.1.2...v2.2.0) (2023-02-08)